refactor: add BusyScope IDisposable pattern and fix untranslated strings
- Add BusyScope() method to PageViewModel for exception-safe busy state management
- Migrate SetBusy/ClearBusy pairs to BusyScope in 7 services:
  SettingsPageService, SavegameCommandService, GameListService,
  GameSearchService, SavegameQueryService, GameWithStatsService,
  GameDetailsService
- Add 8 localization keys (en-US + it-IT): RESTORING_SAVEGAME,
  DELETING_SAVEGAMES, SYNCING_SAVEGAMES, SEARCH_COMPLETED,
  IMPORTING_SAVEGAMES, SAVING_LAUNCH_OPTIONS,
  NO_SAVEGAME_BACKUPS_FOUND, SAVEGAME_BACKUPS_FOUND_IMPORT
- Fix UNINSTALLING_1 bug: use existing UNINSTALLING key with game title

+++ src/TombLauncher/Services/GameWithStatsService.cs
@@ -300,13 +300,18 @@ public class GameWithStatsService : IViewService, IDisposable
     {
         var game = await _gameDataService.GetGameWithStats(gameId);
         var currentPage = NavigationManager.CurrentPage as INavigationTarget;
-        currentPage?.SetBusy("UNINSTALLING_1".GetLocalizedString());
-        var installDir = game.GameMetadata.InstallDirectory;
-        if (installDir != null)
+        if (currentPage is PageViewModel pageVm)
         {
-            Directory.Delete(installDir, true);
+            using (pageVm.BusyScope("UNINSTALLING".GetLocalizedString(game.GameMetadata.Title)))
+            {
+                var installDir = game.GameMetadata.InstallDirectory;
+                if (installDir != null)
+                {
+                    Directory.Delete(installDir, true);
+                }
+                await _gameDataService.MarkGameAsUninstalled(gameId);
+            }
         }
-        await _gameDataService.MarkGameAsUninstalled(gameId);
         await NavigationManager.GoBack();
     }
